requirements.txt
# pip install -r requirements.txt
#to get all dependencies
pandas
yfinance
mysql-connector-python